requirements.txt
# Core app framework
streamlit==1.32.0
fastapi==0.110.0

# PDF parsing
PyPDF2==3.0.1

# Text processing
regex==2023.12.25

# Embeddings
sentence-transformers==2.2.2

# Vector stores
faiss-cpu==1.7.4
chromadb==0.4.22

# LLMs
openai==1.12.0
transformers==4.37.2

# Utilities
numpy==1.26.4
scikit-learn==1.4.0

# Deployment & monitoring (optional)
uvicorn==0.27.0
prometheus-client==0.19.0
sentry-sdk==1.39.1
